pyproject.toml
[project]
name = "api-series"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.3"
dependencies = [
    "cryptography>=46.0.5",
    "fastapi[standard]>=0.135.1",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
]

[tool.fastapi]
entry-point = "app.main:app"
#from app.main import app as app 